refactor(path): changed to absolute paths, changed, db folder to database folder
